+++ src/pages/Articles/LostItemDetailPage/index.tsx
@@ -12,10 +12,6 @@ import { useArticlesLogger } from 'pages/Articles/hooks/useArticlesLogger';
 import DisplayImage from 'pages/Articles/LostItemDetailPage/components/DisplayImage';
 import ReportIcon from 'assets/svg/Articles/report.svg';
 import MessageIcon from 'assets/svg/Articles/message.svg';
-import { useMutation } from '@tanstack/react-query';
-import showToast from 'utils/ts/showToast';
-import useTokenState from 'utils/hooks/state/useTokenState';
-import { reportLostItemArticle } from 'api/articles';
 import ReportModal from 'pages/Articles/LostItemDetailPage/components/ReportModal';
 import styles from './LostItemDetailPage.module.scss';
 
@@ -25,7 +21,7 @@ export default function LostItemDetailPage() {
   const params = useParams();
   const [isDeleteModalOpen, openDeleteModal, closeDeleteModal] = useBooleanState(false);
   const { data: userInfo } = useUser();
-  const token = useTokenState();
+
   const [isReportModalOpen, openReportModal, closeReportModal] = useBooleanState(false);
   // const isCouncil = userInfo && userInfo.student_number === '2022136000';
 
@@ -34,7 +30,6 @@ export default function LostItemDetailPage() {
 
   // const isMyArticle = true; // 내가 쓴 게시글 테스트 용도
   const isMyArticle = userInfo?.student_number === article?.author;
-  console.log(isMyArticle);
 
   const {
     boardId,
@@ -52,38 +47,6 @@ export default function LostItemDetailPage() {
   const handleDeleteButtonClick = () => {
     logFindUserDeleteClick();
     openDeleteModal();
-  };
-
-  // 신고 API 연결
-  const { mutate: reportArticle } = useMutation({
-    mutationFn: (reports: { title: string; content: string }[]) => {
-      if (!token) {
-        showToast('error', '로그인이 필요합니다.');
-        return Promise.reject(new Error('Unauthorized')); // 토큰이 없을 때 에러처리
-      }
-      return reportLostItemArticle(token, articleId, { reports });
-    },
-    onSuccess: () => {
-      showToast('success', '게시글이 신고되었습니다.');
-    },
-    onError: () => {
-      showToast('error', '신고 접수에 실패했습니다.');
-    },
-  });
-
-  const handleReportClick = () => {
-    if (!userInfo) {
-      showToast('error', '로그인이 필요합니다.');
-      return;
-    }
-
-    const reports = [
-      {
-        title: '기타',
-        content: '운영진 확인이 필요한 게시글입니다.',
-      },
-    ];
-    reportArticle(reports);
   };
 
   return (
@@ -114,7 +77,6 @@ export default function LostItemDetailPage() {
             <p>재실 시간은 공지 사항을 참고해 주시기 바랍니다.</p>
           </div>
 
-          {/* "삭제" 버튼 vs "쪽지 보내기 + 신고" 버튼 */}
           <div className={styles['button-container']}>
             {isMyArticle ? (
               <button
@@ -142,12 +104,12 @@ export default function LostItemDetailPage() {
                   <ReportIcon />
                   신고
                 </button>
-                {isReportModalOpen && <ReportModal articleId={articleId} closeReportModal={closeReportModal} />}
+                {isReportModalOpen
+                && <ReportModal articleId={articleId} closeReportModal={closeReportModal} />}
               </>
             )}
           </div>
 
-          {/* 모바일에서는 기존 "목록" 버튼 유지 */}
           <div className={styles.contents__buttons}>
             {isMobile && (
             <button
